+++ src/captchas/DuplicateCaptcha.php
@@ -4,6 +4,7 @@ namespace barrelstrength\sproutforms\captchas;
 
 use barrelstrength\sproutforms\base\Captcha;
 use barrelstrength\sproutforms\events\OnBeforeSaveEntryEvent;
+use barrelstrength\sproutforms\events\OnBeforeValidateEntryEvent;
 use Craft;
 use craft\errors\MissingComponentException;
 
@@ -51,7 +52,7 @@ class DuplicateCaptcha extends Captcha
      * @throws MissingComponentException
      * @throws MissingComponentException
      */
-    public function verifySubmission(OnBeforeSaveEntryEvent $event): bool
+    public function verifySubmission(OnBeforeValidateEntryEvent $event): bool
     {
         $uniqueid = null;